+++ src/App.ts
@@ -59,18 +59,17 @@ export class App {
       .then((r) => r.data as Buffer);
   }
 
-  public async appGet(
-    this: QlikRepoApi,
-    id?: string
-  ): Promise<IApp[] | IAppCondensed[]> {
-    let url = "app";
-    if (id) url += `/${id}`;
+  public async appGet(this: QlikRepoApi, id: string): Promise<IApp> {
+    if (!id) throw new Error(`appGet: "id" parameter is required`);
+    return await this.repoClient
+      .Get(`app/${id}`)
+      .then((res) => res.data as IApp);
+  }
 
-    return await this.repoClient.Get(url).then((res) => {
-      if (!id) return res.data as IAppCondensed[];
-
-      return [res.data] as IApp[];
-    });
+  public async appGetAll(this: QlikRepoApi): Promise<IAppCondensed[]> {
+    return await this.repoClient
+      .Get(`app`)
+      .then((res) => res.data as IAppCondensed[]);
   }
 
   public async appGetFilter(
@@ -221,7 +220,7 @@ export class App {
   public async appUpdate(this: QlikRepoApi, arg: IAppUpdate): Promise<IApp> {
     if (!arg.id) throw new Error(`appUpdate: "id" parameter is required`);
 
-    let app = await this.appGet(arg.id).then((a) => a[0] as IApp);
+    let app = await this.appGet(arg.id);
 
     if (arg.name) app.name = arg.name;
     if (arg.description) app.description = arg.description;