+++ src/ContentLibrary.ts
@@ -47,16 +47,20 @@ export class ContentLibrary {
 
   public async contentLibraryGet(
     this: QlikRepoApi,
-    id?: string
-  ): Promise<IContentLibrary[] | IContentLibraryCondensed[]> {
-    let url = "contentlibrary";
-    if (id) url += `/${id}`;
+    id: string
+  ): Promise<IContentLibrary> {
+    if (!id) throw new Error(`contentLibraryGet: "id" parameter is required`);
+    return await this.repoClient
+      .Get(`contentlibrary/${id}`)
+      .then((res) => res.data as IContentLibrary);
+  }
 
-    return await this.repoClient.Get(url).then((res) => {
-      if (!id) return res.data as IContentLibraryCondensed[];
-
-      return [res.data] as IContentLibrary[];
-    });
+  public async contentLibraryGetAll(
+    this: QlikRepoApi
+  ): Promise<IContentLibraryCondensed[]> {
+    return await this.repoClient
+      .Get(`contentlibrary`)
+      .then((res) => res.data as IContentLibraryCondensed[]);
   }
 
   public async contentLibraryGetFilter(
@@ -225,9 +229,7 @@ export class ContentLibrary {
     if (!arg.id)
       throw new Error(`contentLibraryUpdate: "id" parameter is required`);
 
-    let contentLibrary = await this.contentLibraryGet(arg.id).then(
-      (c) => c[0] as IContentLibrary
-    );
+    let contentLibrary = await this.contentLibraryGet(arg.id);
 
     let updateCommon = new UpdateCommonProperties(this, contentLibrary, arg);
     contentLibrary = await updateCommon.updateAll();