+++ src/User.ts
@@ -15,18 +15,17 @@ import { IUserUpdate, IUserCreate } from "./interfaces/argument.interface";
 export class User {
   constructor() {}
 
-  public async userGet(
-    this: QlikRepoApi,
-    id?: string
-  ): Promise<IUser[] | IUserCondensed[]> {
-    let url = "user";
-    if (id) url += `/${id}`;
+  public async userGet(this: QlikRepoApi, id: string): Promise<IUser> {
+    if (!id) throw new Error(`userGet: "id" parameter is required`);
+    return await this.repoClient
+      .Get(`user/${id}`)
+      .then((res) => res.data as IUser);
+  }
 
-    return await this.repoClient.Get(url).then((res) => {
-      if (!id) return res.data as IUserCondensed[];
-
-      return [res.data] as IUser[];
-    });
+  public async userGetAll(this: QlikRepoApi): Promise<IUserCondensed[]> {
+    return await this.repoClient
+      .Get(`user`)
+      .then((res) => res.data as IUserCondensed[]);
   }
 
   public async userGetFilter(
@@ -103,7 +102,7 @@ export class User {
   public async userUpdate(this: QlikRepoApi, arg: IUserUpdate): Promise<IUser> {
     if (!arg.id) throw new Error(`userUpdate: "id" parameter is required`);
 
-    let user = await this.userGet(arg.id).then((u) => u[0] as IUser);
+    let user = await this.userGet(arg.id);
 
     if (arg.roles) user.roles = arg.roles;
     if (arg.name) user.name = arg.name;